+++ src/align-versions.js
@@ -10,7 +10,7 @@ import { isMonorepoRoot, getSubProjectDirectories, pkg } from './utils.js'
 /**
  * @typedef {import("./types.js").GlobalOptions} GlobalOptions
  * @typedef {import("./types.js").ReleaseOptions} ReleaseOptions
- * @typedef {import("listr").ListrTaskWrapper} Task
+ * @typedef {import('listr').ListrTaskWrapper} Task
  */
 
 const tasks = new Listr([