+++ src/test/browser.js
@@ -3,6 +3,7 @@ import { fileURLToPath } from 'url'
 import { execa } from 'execa'
 import merge from '../utils/merge-options.js'
 import { fromAegir, findBinary } from '../utils.js'
+import { killIfCoverageHangs } from './utils.js'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
@@ -41,7 +42,7 @@ export default async (argv, execaOptions) => {
   const beforeEnv = before && before.env ? before.env : {}
 
   // run pw-test
-  await execa(findBinary('pw-test'),
+  const proc = execa(findBinary('pw-test'),
     [
       ...files,
       '--mode', argv.runner === 'browser' ? 'main' : 'worker',
@@ -68,6 +69,8 @@ export default async (argv, execaOptions) => {
       execaOptions
     )
   )
+
+  await killIfCoverageHangs(proc, argv)
 
   // after hook
   await argv.fileConfig.test.after(argv, before)