+++ test/fixtures/dependency-check/jsx-fail/src/index.jsx
@@ -1,9 +1,9 @@
 // @ts-expect-error - not installed
-// eslint-disable-next-line no-unused-vars
+
 import React from 'react'
 
 // @ts-expect-error - not installed
-// eslint-disable-next-line no-unused-vars
+
 const Component = () => (<div>Hello, world!</div>)
 const App = () => (<Component />)