+++ CHANGELOG.md
@@ -1,3 +1,95 @@
+## [47.1.3](https://github.com/ipfs/aegir/compare/v47.1.2...v47.1.3) (2026-04-10)
+
+### Bug Fixes
+
+* revert "deps: bump @types/node from 22.19.7 to 25.2.0" ([#1935](https://github.com/ipfs/aegir/issues/1935)) ([7ab36a4](https://github.com/ipfs/aegir/commit/7ab36a464ac3bae4f853af577fb702f6692a794b))
+
+## [47.1.2](https://github.com/ipfs/aegir/compare/v47.1.1...v47.1.2) (2026-04-10)
+
+### Bug Fixes
+
+* update unknown type resolver ([15ad88d](https://github.com/ipfs/aegir/commit/15ad88d3da082fed61485ae74466817df602e4ec))
+
+## [47.1.1](https://github.com/ipfs/aegir/compare/v47.1.0...v47.1.1) (2026-04-10)
+
+### Bug Fixes
+
+* do not replace - with _ in module name ([750d2e8](https://github.com/ipfs/aegir/commit/750d2e8b98d2e5a2e54d1f58fc1a36136d119f17))
+
+## [47.1.0](https://github.com/ipfs/aegir/compare/v47.0.26...v47.1.0) (2026-04-10)
+
+### Features
+
+* add module-sync to default package config ([#1921](https://github.com/ipfs/aegir/issues/1921)) ([c7cc013](https://github.com/ipfs/aegir/commit/c7cc013b8de3b3a718bc960693f70f59f489e29b))
+* search for .aegir.mjs config file ([#1901](https://github.com/ipfs/aegir/issues/1901)) ([a6c4790](https://github.com/ipfs/aegir/commit/a6c47901b4efbbb42ec703a5bf423d134fe38e7a))
+
+### Bug Fixes
+
+* include link to NodeJS.CallSite in docs ([155ff85](https://github.com/ipfs/aegir/commit/155ff8507361d119bd2792f1d918303600f6bb4f))
+* print stack trace on error ([#1852](https://github.com/ipfs/aegir/issues/1852)) ([387c8d2](https://github.com/ipfs/aegir/commit/387c8d24c961665d68c6695b79d518e02fa73746))
+* update default build files ([#1922](https://github.com/ipfs/aegir/issues/1922)) ([6c4ee02](https://github.com/ipfs/aegir/commit/6c4ee027179a06bf91d2af3363a640f95dd55639))
+
+### Trivial Changes
+
+* fix typo in comment ([#1920](https://github.com/ipfs/aegir/issues/1920)) ([194d9ed](https://github.com/ipfs/aegir/commit/194d9ed232a24271ae709b85aa79d7522ea7522f))
+* lint actions ([#1924](https://github.com/ipfs/aegir/issues/1924)) ([5372055](https://github.com/ipfs/aegir/commit/5372055d52bf97a2b6a252a7b726570af7aa8edd))
+
+### Documentation
+
+* fix link to types in readme ([#1923](https://github.com/ipfs/aegir/issues/1923)) ([b7681c4](https://github.com/ipfs/aegir/commit/b7681c45c400df04fc4836b7c53c2b9e92b9429f))
+
+### Dependencies
+
+* bump @types/node from 22.19.7 to 25.2.0 ([#1919](https://github.com/ipfs/aegir/issues/1919)) ([6a7a09d](https://github.com/ipfs/aegir/commit/6a7a09d4d5f1273d7de87b890354dfb3c2ac2a95))
+* bump env-paths from 3.0.0 to 4.0.0 ([#1916](https://github.com/ipfs/aegir/issues/1916)) ([bf19b4d](https://github.com/ipfs/aegir/commit/bf19b4d95c36da2855e0dd5bc5cd8407f63ea465))
+* bump esbuild from 0.25.12 to 0.27.2 ([#1908](https://github.com/ipfs/aegir/issues/1908)) ([82dac8c](https://github.com/ipfs/aegir/commit/82dac8cdc776e19faf1b4d310eb77f94548d362f))
+* bump globby from 15.0.0 to 16.1.0 ([#1909](https://github.com/ipfs/aegir/issues/1909)) ([4e5aa85](https://github.com/ipfs/aegir/commit/4e5aa85578804e6db3320e3d0500bd65e795caae))
+* bump the eslint group across 1 directory with 2 updates ([#1880](https://github.com/ipfs/aegir/issues/1880)) ([be25ad7](https://github.com/ipfs/aegir/commit/be25ad7d0c6014d47bc815b606eb45720680649b))
+* **dev:** bump electron from 38.8.0 to 40.1.0 ([#1918](https://github.com/ipfs/aegir/issues/1918)) ([9bc6d14](https://github.com/ipfs/aegir/commit/9bc6d1411f0a1edaa3554d143be6e507570b6182))
+
+## [47.0.26](https://github.com/ipfs/aegir/compare/v47.0.25...v47.0.26) (2026-01-22)
+
+### Dependencies
+
+* bump semantic-release to 25.0.2 ([#1904](https://github.com/ipfs/aegir/issues/1904)) ([a628685](https://github.com/ipfs/aegir/commit/a628685d00bc3ade832b295048871a120036381b))
+
+## [47.0.25](https://github.com/ipfs/aegir/compare/v47.0.24...v47.0.25) (2025-12-07)
+
+### Bug Fixes
+
+* remove forced test exit on windows ([#1897](https://github.com/ipfs/aegir/issues/1897)) ([0ca62ab](https://github.com/ipfs/aegir/commit/0ca62ab47fe4120c332f6c0f59bb78faa174c00f))
+
+## [47.0.24](https://github.com/ipfs/aegir/compare/v47.0.23...v47.0.24) (2025-10-17)
+
+### Dependencies
+
+* bump typescript-docs-verifier from 2.5.3 to 3.0.1 ([#1841](https://github.com/ipfs/aegir/issues/1841)) ([493cd48](https://github.com/ipfs/aegir/commit/493cd4864487e73c96f13952b895c4841d469da0))
+
+## [47.0.23](https://github.com/ipfs/aegir/compare/v47.0.22...v47.0.23) (2025-10-14)
+
+### Dependencies
+
+* bump globby from 14.1.0 to 15.0.0 ([#1867](https://github.com/ipfs/aegir/issues/1867)) ([583d846](https://github.com/ipfs/aegir/commit/583d846486b23a9b8f95c11a422f8aec68dcff20))
+* bump p-queue from 8.1.1 to 9.0.0 ([#1871](https://github.com/ipfs/aegir/issues/1871)) ([3dec148](https://github.com/ipfs/aegir/commit/3dec148e40c5251505eb20d4c868031466ecc2e7))
+
+## [47.0.22](https://github.com/ipfs/aegir/compare/v47.0.21...v47.0.22) (2025-09-17)
+
+### Bug Fixes
+
+* time out browser coverage collection ([#1865](https://github.com/ipfs/aegir/issues/1865)) ([16ba8c7](https://github.com/ipfs/aegir/commit/16ba8c7bd4e3a2e17b470022efe3183a14f9090d))
+
+### Dependencies
+
+* bump npm-package-json-lint from 8.0.0 to 9.0.0 ([#1827](https://github.com/ipfs/aegir/issues/1827)) ([94bda99](https://github.com/ipfs/aegir/commit/94bda99e3e2c907a1753e9b5c97336ee9f313ce6))
+* bump p-retry from 6.2.1 to 7.0.0 ([#1856](https://github.com/ipfs/aegir/issues/1856)) ([b578e1d](https://github.com/ipfs/aegir/commit/b578e1df0fa65fab28d078123cd2249e93f9c725))
+* **dev:** bump electron from 36.9.1 to 38.1.1 ([#1864](https://github.com/ipfs/aegir/issues/1864)) ([27c911b](https://github.com/ipfs/aegir/commit/27c911b8def49ba8847b1f50f85d77186b5456fa))
+
+## [47.0.21](https://github.com/ipfs/aegir/compare/v47.0.20...v47.0.21) (2025-08-05)
+
+### Bug Fixes
+
+* pin typescript ([#1846](https://github.com/ipfs/aegir/issues/1846)) ([b5f0e94](https://github.com/ipfs/aegir/commit/b5f0e9498fb3cf5fec20c2f95d14f47807cad2f7))
+
 ## [47.0.20](https://github.com/ipfs/aegir/compare/v47.0.19...v47.0.20) (2025-07-07)
 
 ### Bug Fixes